requirements.txt
rasa==3.6.2
rasa-sdk==3.6.0
pymysql==1.0.3
python-dotenv==1.0.0